note: split store.go and storage.go into store/query/slug/tags
